requirements.txt
# KMS 2.5 Vector Search - Python Dependencies

# ============================================================================
# Core Web Framework
# ============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ============================================================================
# Authentication & Security
# ============================================================================
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# ============================================================================
# Data Processing
# ============================================================================
pyspark==3.4.1
pandas==2.1.3
numpy==1.26.2

# ============================================================================
# Vector Database
# ============================================================================
weaviate-client==3.25.0

# ============================================================================
# NLP & Embeddings
# ============================================================================
spacy==3.7.2
# Download model: python -m spacy download en_core_web_sm
transformers==4.35.0
torch==2.1.0  # For transformers if needed

# ============================================================================
# PII Detection
# ============================================================================
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33

# ============================================================================
# Apache Airflow (for pipeline orchestration)
# ============================================================================
apache-airflow==2.7.3
apache-airflow-providers-http==4.5.1

# ============================================================================
# Database
# ============================================================================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # For PostgreSQL (Airflow)

# ============================================================================
# Monitoring & Observability
# ============================================================================
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# ============================================================================
# HTTP Client
# ============================================================================
httpx==0.25.1
requests==2.31.0

# ============================================================================
# Data Validation
# ============================================================================
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# ============================================================================
# Configuration & Environment
# ============================================================================
python-dotenv==1.0.0

# ============================================================================
# Utilities
# ============================================================================
python-dateutil==2.8.2
pytz==2023.3

# ============================================================================
# Testing
# ============================================================================
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.1  # For testing async APIs
locust==2.17.0  # For load testing

# ============================================================================
# Code Quality
# ============================================================================
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# ============================================================================
# Logging
# ============================================================================
loguru==0.7.2
